+++ src/IoTSharp.Data.Taos/Protocols/TDRESTful/TaosRESTful.cs
@@ -170,18 +170,4 @@ namespace IoTSharp.Data.Taos.Protocols.TDRESTful
             throw new NotSupportedException("RESTful  不支持 ExecuteBulkInsert");
         }
     }
-
-    public class TaosResult
-    {
-        public int code { get; set; }
-        public string desc { get; set; }
-
-        public List<List<string>> column_meta { get; set; }
-
-        public JArray data { get; set; }
-
-        public int rows { get; set; }
-
-        public object Scalar => (data?.First?.First as JValue)?.Value;
-    }
 }
