+++ src/IoTSharp.Data.Taos/Protocols/TDWebSocket/WSActionReq.cs
@@ -1,0 +1,16 @@
+﻿namespace IoTSharp.Data.Taos.Protocols.TDWebSocket
+{
+    public class WSActionReq<T>
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        public string action { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public T args { get; set; }
+    }
+
+
+}
