+++ src/IoTSharp.Data.Taos/Protocols/TDWebSocket/conn.cs
@@ -1,0 +1,24 @@
+﻿namespace IoTSharp.Data.Taos.Protocols.TDWebSocket
+{
+    public class conn
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        public int req_id { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string user { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string password { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string db { get; set; }
+    }
+
+
+}
